backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
distro==1.9.0
fastapi==0.135.2
firebase_admin==7.4.0
flake8==7.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.4.1
google-crc32c==1.8.0
google-genai==1.69.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.2.0
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.9
pytokens==0.4.1
requests==2.33.1
requests-oauthlib==2.0.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
